+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - continuation prompt based on context.
 - missing type support for increase compatibility.
 - `missing` function: Create missing values.
+- `renameStructField`: Rename field in a structure.
 
 ### Changed